nio: (std,cache,bindc) [SEVERE] registration check for path.
As far as I checked, a new unit connection for already opened file
fails using tnb_file_is_nio(), with GNU fortran/gcc (Debian 14.1.0-5
or 14.1.0).  On the other hand, cache_open_read() seems to accept
this feature.  Actually this feature is processor dependent.
Fortran standard mentions that:
     If a file is already connected to a unit, an OPEN statement
     on that file with a different unit shall not be executed.
Thus whether or not succeeds is unexpected.
Workarounds are introduced to bypass this feature, when a file is
already connected and registered into Nio cache table.

Strict file management is required in this case.  If unit management is
done outside, then it is possible to have same io unit for different
file in cache.  At the moment the users must take care of it, to avoid
external unit management.