nio: (cache) [FATAL] introduced logical-shape reading.
Argument START was not well-defined in the implementation.
Now START is defined default as relative position along a
coordinate (physical storage in a sence).  Thus start=0 corresponds to
the slice start position whenever ISTRn is not defined as 0.
You can control this behavier to call with addition optional argument
blog as TRUE.  Again, FALSE is default for compatibility.