zbt: (plot) [FATAL] Fix binding of sliced array-file.
Arrays in NIO-file with non-default slices (header attributes of
ISTRn, IENDn) do not work as expected.

Indeed, there is confusion whether or not the argument start is
relative to ISTRn, which are have not been defined well.
New property _TouzaCoreVar.pdims is introduced, which corresponds to
/physical/ extents in the file to deal with this confusion.  Argument
start is now defined as relative value to the ISTRn position.

For normal variable of sliced array, extent and pdims properties are
identical,  while those of coordinate variables assigned internally
are different.

----------------------------------------------------------------------
Array of logical shape (0:12, 10:40, 60:150, 100:250):

[DEBUG] <zbt.dsnio> <b'UO'> e=0 record 12 (0, 12) (0, 12)
[DEBUG] <zbt.dsnio> <b'UO'> e=13 b'OCDEPT63' 30 (10, 40) (10, 40)
[DEBUG] <zbt.dsnio> <b'UO'> e=slice(None, None, None) b'OCLATTPV256' 90 (60, 150) (60, 150)
[DEBUG] <zbt.dsnio> <b'UO'> e=slice(None, None, None) b'OCLONTPV360' 150 (100, 250) (100, 250)
----------------------------------------------------------------------
And their coordinates:
[DEBUG] <zbt.dsnio> <b'OCLONTPV360'> e=0 b'OCLONTPV360' 240 (60, 300) (0, 360)
[DEBUG] <zbt.dsnio> <b'OCLATTPV256'> e=0 b'OCLONTPV360' 240 (60, 300) (0, 360)

Therefore, from user points of view, array slicing is performed not
with begin:end but 0:size region.