zbt: (control) [FATAL] remove reuse of Normalize.
From matplotlib 3.10 or so, Normalize.autoscale() in adjust_norm()
method does not work well at slice turning, which raise an error of
NotImplementedError('cannot remove artist').

From deep debuging, Normalize instance for colorbar or color plotting
is a key for this error.  It seems that a kind of double-free of
artist instances happens.

  bar: <class 'matplotlib.colorbar.Colorbar'> 139830507924384
  <class 'matplotlib.collections.QuadMesh'> 139830218183120
  NotImplementedError: cannot remove artist <class 'matplotlib.collections.QuadMesh'> 139830218183120

At definition of a new norm, _remove_method() is bound, however
somewhere in the execution it disappears.

From trial and error, commenting out of norm.autoscale helps to
prohibit raising of the error:

   if saved:                   # cached Normalize instance
      norm = saved
      # norm.autoscale(lims)   #  THIS LINE