std: (choice, etc) [FATAL] updated legacy elemental designs.
Gave up elemental function.

Elemental choice function family does not perfectly work on some
processors.

  integer,intent(in),optional :: B(:)
  ! .....
  A(:) = choice(0, B)          ! case 0
  A(:) = choice(X(:), B)       ! case 1

If B is not present, then case 0 raise a segmentation fault.
Indeed, the style of case 0 is not accepted by Fortran 2018 standard
at least.

  3 An optional dummy argument that is not present is subject to the
  following restrictions.
  :
  (6) If it is an array, it shall not be supplied as an actual
  argument to an elemental procedure unless an array of the same rank
  is supplied as an actual argument corresponding to a nonoptional
  dummy argument of that elemental procedure.

Regarding this limitation, case 2 should work because the nonoptional
argument is supplied.  However, at least on flang (11.0.0) or
nvfortran (24.7.0), it also raises a segmentation fault.
On ifx (2024.0.2), the situation is even complex.  If the module and
caller procedures both are written in a same file, it works (!?),
and if they are separated into two files, it fails.

Whatever the compiler should be, an elemental procedure with optional
argument may not perform as expected, and also it is not easy to
detect the failure at compilation times.  Therefore, I gave up to
implement the elemental choice functions.

Instead, choice procedures to accept array argument(s) are implemented
separately, which will be helpful for some cases.
Consequetly, after some trials, procedures achoice schoice are
implemented as subroutines, instead of fucntions. Function to return
arbitraly-size array may not work....