trp: (float) [FATAL] OPT_TRAPICHE_INTEL_SPECIALS to control ANINT().
Intel compiler aborts with segmentation fault when with -xHost
argument, which is found using Intel_oneAPI 2024 on ES4 SX Aurora.

After some trials, calling ANINT() is the key for the failure.
It is strange but, it does succeed when ANINT() is replaced by a
comparable function, using AINT().

You can switch to the original procedure to define
OPT_TRAPICHE_INTEL_SPECIALS as 0.  In the current implementation,
this macro is set as 1 when __INTEL_COMPILER is non-zero.